# Change `nn.Embedding` from matmul to lookup