@chenyuxyz 2023-12-14
this one is interesting. __getitem__ also uses arange, so the complexity is the same and it's not doing a lookup.

this adds a lot more kernels (see the CI failure), but it's different and faster compared to master due to (1) we are doing a dtype refactor + __getitem__ refactor now, and the __getitem__ uses dtype int for index buffer, while the old code path uses float to pass index (2) the memory access pattern changed

relevant kernel

master
```
    float val0 = *(data1+(gidx0*1608224)+(lidx2*201028)+ridx0);
    float val1 = *(data1+(gidx0*1608224)+(lidx2*201028)+ridx0+50257);
    float val2 = *(data1+(gidx0*1608224)+(lidx2*201028)+ridx0+100514);
    float val3 = *(data1+(gidx0*1608224)+(lidx2*201028)+ridx0+150771);
```

this pr
```
    int val0 = *(data1+(gidx0*32)+(lidx2*4)+(ridx0*4096));
    int val1 = *(data1+(gidx0*32)+(lidx2*4)+(ridx0*4096)+1);
    int val2 = *(data1+(gidx0*32)+(lidx2*4)+(ridx0*4096)+2);
    int val3 = *(data1+(gidx0*32)+(lidx2*4)+(ridx0*4096)+3);
```

since it adds a few kernels it's not a clear win. we can merge this approach if you can make it with the same number of kernels. or you can solve the O(n) embedding bounty that remove the 50257 from reduce idx