@mcognetta 2023-12-15
> since it adds a few kernels it's not a clear win. we can merge this approach if you can make it with the same number of kernels. or you can solve the O(n) embedding bounty that remove the 50257 from reduce idx

By this you mean something like adding a file akin to extra/gemm/cuda_matmul.py but for an embedding table lookup?