@chenyuxyz 2023-12-16
no the embedding bounty needs to be merged to core tinygrad and make it the default behavior