@chenyuxyz 2024-01-09
closing this as stale. feel free to open a new one with more rigorous perf tests. even though it's not the same as the Embedding bounty, we are interested in perf improvements in general